+++ lib/flutter_cached.dart
@@ -1,4 +1,5 @@
 library flutter_cached;
 
 export 'src/controller.dart';
+export 'src/update.dart';
 export 'src/widgets.dart';